requirements.txt
crewai>=0.80.0
crewai-tools>=0.10.0
python-dotenv>=1.0.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-ollama>=0.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.1
neo4j>=5.19.0
httpx>=0.27.0
pandas>=2.2.0
scikit-learn>=1.4.0
dowhy>=0.11
pytest>=8.2.0
bcrypt>=4.1.2

